Introduce FailureHookRunnerUseCase with DI
Refactor the failure-hook logic into a testable, dependency-injected use case and wire a tiny production shim. FailureHookRunner now delegates to FailureHookRunnerUseCase (moved business logic there) which accepts ScopePreferencesStoreProtocol and TerminalLauncherProtocol; both protocols were added to RunnerBarCore. Added DefaultScopePreferencesStore and DefaultTerminalLauncher adapters to bridge the existing singletons, preserving behavior: branch filtering, token resolution, fetching failed job logs, Task.detached for background work, and MainActor hop for launching Terminal. This change keeps the existing public API (FailureHookRunner.fireIfNeeded) while making the core logic unit-testable and easier to mock.
